requirements.txt
Flask==2.0.1
Dash==2.0.0
plotly==5.3.1
Pillow==8.3.1
Faker==8.10.0
pandas==1.3.3
flask-cors==3.0.10
